fix ballot strategies not translating
