+++ src/constants/ballotStrategies/get-ballot-strategies-by-address.ts
@@ -1,0 +1,10 @@
+import { ballotStrategies, customStrategy } from './ballot-strategies'
+
+// Put in separate files because lingui.js t macro was not working on ballot strategies
+export const getBallotStrategyByAddress = (address: string) => {
+  const s =
+    ballotStrategies().find(
+      s => s.address.toLowerCase() === address.toLowerCase(),
+    ) ?? customStrategy(address)
+  return s
+}
